Support for special "hint" notes.
The hint notes are to be used in conjunction with variable band positions and references, to prevent the randomizer from picking the band they reside on.
Other than that, they are invisible and don't have any effect on the gameplay. By default, they are removed from the beatmap entirely after variable bands and references are applied. Beatmap creators may wish to use game.strip-hintnotes=false and frontend.pygame.gamerenderer.show-hintnotes=true to see where the hint notes are.